Remove the default publishing box and start a custom one
We should be able to better guide the overall publishing experience
for a profile and can exclude things like scheduling a post, viewing
revisions (for now), and such.

This commit removes the default submit area and starts to replace
it with a skeleton of its former self. Next we'll try to move the
Network ID loading process into this area.