Add support for REST API requests by wsu_nid
Previously, we supported the open querying of meta_key and
meta_value. In the release of the REST API in WordPress 4.7, the
filter allowing this changed a bit, forcing us to adjust.

Rather than duplicate the behavior, this gives us the opportunity
to restrict things a bit and only allow queries by a whitelist of
query vars.

This adds support for the wsu_nid query parameter to REST API
requests on the people post type.