Break up the main class into a couple separate smaller classes
In preparation of consolidating the People Display functionality
into this plugin.

Aside from adjustments to the hooks, variable references, one or
two function names, and docblocks, this is just a reorganization
of the existing code. Well, that's not entirely true - I did drop
the code for removing some image sizes and all of the capabilities
related code. Capabilities will be addressed later.