Fix manual refresh handling
This fixes the AJAX call that I previously broke, but perhaps more
importantly, refacts the JavaScript to avoid duplicating the
repeatable fields and photos when refreshing a profile with data
from people.wsu.edu.

Though there's likely a more elegant way to write some of these,
this works and will at least serve to get the idea(s) across.