Avoid saving a NID to a page if REST API request fails
I ran into an issue where a REST API request for my NID failed, but
the NID was still saved to the directory page. This prevented the
profile from being stored on the local site, which confused things
when the directory page was displayed.

We can work around this by only storing NIDs that come back with
successful REST API data or have already been created as profiles
on the local site.