Split the "Edit Profile" JavaScript into two files
This might be madness, but it seemed like the right course
yesterday afternoon, so I'm going to stick to it.

The code in `admin-edit-profile-secondary.js` is only needed on
sites other than the primary directory - even though some of its
functions are referenced in `admin-edit-profile.js`. But, those
functions will only be called on sites other than the primary
directory.

I've occasionally seen an error where the script is trying to
`setContent` for the editors before TinyMCE is initialized, which
informed the decision to enqueue `admin-edit-profile-secondary.js`
from within `wp_enqueue_editor`. I haven't seen it since, so I'm
hoping this has made it more reliable.

The `secondary` script checks for changes before posting back to
people.wsu.edu. This is probably overkill, but it seemed like a
good idea at the time.

Lastly, this is the start of an effort to (re)name JavaScript
files in a way that better indicates where they're used.