Implement a proof of concept permissions check
This is pretty bad, but it gets the general idea across. Even if a
user doesn't have sufficient permissions to edit a profile, they
should still be able to select which Working Title, Biography, and
Photo to display on the front end for their particular site. So,
this demonstrates how that could work - though I'm sure there's a
nicer way to do it. Like one that doesn't throw a console error.

I'm also wondering if doing it the other way around would make
more sense (that is, disable all fields by default and only enable
them if a user has permissions to edit the profile).