Leverage primary record post IDs to make a single REST request
Previously, I looped through the local records and made a request
for each one - which typically failed to load all the people for a
given page, and likely would never have been appropriate for
production.

There's a possibility that I missed a thing or two with this
transition, but things seem to be working as expected so far.