Retrieve person data from an organizational source
This allows the filtering of person data retrieval from NID so that
other organizations outside of WSU have a better chance of using
the plugin.

This also allows us to simplify the code a bit because we can move
the normalization of our AD data into the central theme. The side
effect of this being that NID lookups will fail by default on non
central sites. We'll need to figure out if this has any side effects.