Avoid a race condition when populating editor instances
Depending on the speed of the REST API connection, it's entirely
possible that the data for a person is available before TinyMCE
has fully initialized the editors. In this case, a fatal JS error
will appear because the editors are not available to be populated.

We can hook into the TinyMCE initialization and provide our own
callback when each editor is complete. We can then flag when the
data from the REST API is available and has been processed enough
to start populating.

This starts to refactor the JS a bit to live in the wsuwp.people
namespace so that we may have clear ways to share it across scripts
in the future.