Update the method for keeping a person's `listed_on` data in sync
This is still probably too rudimentary for production, as we need
a way to loop through all profiles added to a directory page and
update their `listed_on` data. This currently relies on the `edit`
button being clicked to trigger that update.

However, this commit introduces the removal of a `listed_on` entry
when a profile is removed from a directory page, which I think _is_
sufficient for production (though I could be wrong).