Rework editor interface markup to better match person card markup
Step one :). Things got painful for a bit and the execution might
be a bit strange in places, but things are very close to matching
now. One exception is that each Degree and Title is in its own
`span` wrapper, and I think it makes sense to make that change on
to the person card markup, too, if we need to have a perfect match.