Use the `single.php` template for individual profile views
This adds some elements that we may not particularly want, but is
consistent with expected behavior from a non-heirarchical post
type.

We may need to provide a `articles/post-wsuwp_people_profile.php`
template - which is what I ended up needing to do in the Murrow
theme - in the Spine theme to start us out with a cleaner slate.