Account for usage of the `wsuwp_people_default_rewrite_slug` filter
If the filter is being used, it seems safe to assume that all
profiles will be expected to at the `{slug}/{profile-slug}/` path.

Additionally, the added rewrite rules don't appear to have any
effect when the filter is used - which I think makes sense -  so
we can skip all that stuff in those cases.