Update directory template filters CSS
This uses the same styles shipped with Content Syndicate People.
The idea is to provide an accessible foundation that isn't too
difficult to override if needed.

This commit also renames the associated CSS files. In the future,
this (and the JS) might be broken into a few files that are enqueued
as needed.